move `constraint_programming`
